requirements.txt
langgraph
langchain_ollama
langchain_tools
langchain_community
elasticsearch<9